split core / rxjava2 library
