+++ example/build.gradle
@@ -18,7 +18,6 @@ android {
       proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
     }
   }
-  productFlavors {}
 }
 
 dependencies {
@@ -30,6 +29,6 @@ dependencies {
   androidTestImplementation 'com.android.support.test:runner:1.0.1'
   androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
 
-  implementation project(':library')
+  implementation project(':core')
   implementation "io.reactivex.rxjava2:rxjava:2.1.10"
 }
