+++ example/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.yuki312.rxbackoff.example"
+  package="com.yuki312.backoff.example"
   >
 
   <application
@@ -11,7 +11,7 @@
     android:supportsRtl="true"
     android:theme="@style/AppTheme"
     >
-    <activity android:name=".MainActivity">
+    <activity android:name="com.yuki312.backoff.example.MainActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
