+++ example/src/main/java/com/yuki312/backoff/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.yuki312.backoff.example
+
+import android.support.v7.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity()
