--- example/src/main/java/com/yuki312/rxbackoff/example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.yuki312.rxbackoff.example
-
-import android.support.v7.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity()
