+++ gradle.properties
@@ -15,8 +15,32 @@ org.gradle.jvmargs=-Xmx1536m
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
-VERSION_NAME=0.0.2
-VERSION_CODE=2
+
+COMPILE_SDK_VERSION=27
+TARGET_SDK_VERSION=27
+MIN_SDK_VERSION=16
+
+
+# # # # # # # # # # # # # # # #
+# LIBRARIES
+# # # # # # # # # # # # # # # #
+SUPPORT_LIB=27.1.0
+MOCKITO_CORE=2.15.0
+ASSERTJ_CORE=3.9.0
+JUNIT=4.12
+
+
+# # # # # # # # # # # # # # # #
+# BINTRAY
+# # # # # # # # # # # # # # # #
 USER_ORG=yuki312
 GROUP_ID=com.yuki312
-ARTIFACT_ID=RxBackoff
+WEBSITE=https://github.com/YukiMatsumura/RxBackoff
+
+CORE_VERSION_NAME=0.0.2
+CORE_VERSION_CODE=2
+CORE_ARTIFACT_ID=backoff-core
+
+RXJAVA2_VERSION_NAME=0.0.2
+RXJAVA2_VERSION_CODE=2
+RXJAVA2_ARTIFACT_ID=backoff-rxjava2
