--- library/build.gradle
@@ -1,54 +1,0 @@
-apply plugin: 'com.android.library'
-apply from: '../jacoco.gradle'
-apply plugin: 'com.novoda.bintray-release'
-
-android {
-  compileSdkVersion 27
-
-  defaultConfig {
-    minSdkVersion 19
-    targetSdkVersion 27
-    versionCode VERSION_CODE as int
-    versionName VERSION_NAME
-
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-  }
-
-  buildTypes {
-    debug {
-      testCoverageEnabled true
-    }
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_7
-    targetCompatibility JavaVersion.VERSION_1_7
-  }
-
-  lintOptions {
-    textReport true
-  }
-}
-
-dependencies {
-  implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-  implementation 'com.android.support:appcompat-v7:27.1.0'
-  testImplementation 'junit:junit:4.12'
-  androidTestImplementation 'com.android.support.test:runner:1.0.1'
-  androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
-
-  api "io.reactivex.rxjava2:rxjava:2.1.10"
-  testImplementation 'org.assertj:assertj-core:3.9.0'
-  testImplementation "org.mockito:mockito-core:2.15.0"
-}
-
-// ./gradlew clean :library:build :library:bintrayUpload -PbintrayUser=[bintrayUser] -PbintrayKey=[bintrayKey] -PdryRun=false
-publish {
-  userOrg = USER_ORG
-  groupId = GROUP_ID
-  artifactId = ARTIFACT_ID
-  publishVersion = VERSION_NAME
-  desc = "Backoff algorithm with RxJava2"
-  website = "https://github.com/YukiMatsumura/RxBackoff"
-}
