+++ rxjava2/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.yuki312.backoff.rxjava2"/>
