+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':example', ':library'
+include ':example', ':core', ':rxjava2'
